Move circular velocity to potential
This also separates halo and gas_halo into two files (halo and halo_set, gas_halo and gas_haloset). Since the set files rely on functions in potential but functions in potential rely on functions in halo/gas_halo. In the future, might change this to halo -> halo_util and halo_set -> halo